+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.0] - 2024-09-15
+
+### Fixes
+
+* Repairs to BASIC renumbering
+    - updates all references
+    - moves can handle blank lines
+
+### New Behaviors
+
+* The `LineNumberTool` trait is deprecated, use the `Renumber` trait instead.
+* Improved LSP symbol manipulations and checks
+* LSP declarations, definitions, and references do not overlap.
+
 ## [3.1.0] - 2024-09-08
 
 ### New Features